fix(functor-lang): tighten the unknown-member suggestion; consolidate the drift lock
From the cross-engine review (both engines flagged both items):

- nearest_member only suggests an UNAMBIGUOUS prefix match, and only in the
  "you typed a prefix of the real name" direction. `List.f` (filter/flatten/
  fold) and `Math.a` (abs/atan2) now list the namespace instead of naming an
  arbitrary member, and `List.mapIndexed` no longer answers "did you mean
  List.map?" — a function with different semantics. Math.clamp01, Debug.log
  and Math.sine suggestions are unaffected.
- Drop the source-scraping drift test: complete.rs already had a STRONGER
  compile-time guard (an exhaustive match over Builtin) over the same const.
  Move it beside ALL_BUILTINS where the const now lives, and add the one
  check it lacked — a duplicate entry masking a missing builtin.
- Drop a tautological assert, fix stale test names in a doc comment, and note
  that Random is closed by the module system, not by this diagnostic.
